Renamed two route files, for consistency with learnforge tutorial.
